DATA_PROJECTS/demos_gegen_rechtsextremismus/requirements.txt
requests
pandas
matplotlib
numpy
folium
opencage
python-dotenv